S-9: scheduler god class에서 EventShadowManager 분리
P2 2-4 event-driven shadow 구독/저널링 로직(~327줄)을 StrategyScheduler에서
scheduler/event_shadow_manager.py의 EventShadowManager로 추출한다.

- entry/exit router 구독 갱신, evaluator wrapper, shadow journal record/flush,
  카테고리 키 헬퍼, 전략 비활성화 teardown을 manager로 이동.
- StrategyScheduler는 manager를 생성·위임만 한다. 공유 협력자(market_clock,
  price_subscription_service, get_strategy_holdings 콜백)는 주입, shadow 전용
  상태(event_router, journal, entry/exit 구독 set)는 manager가 소유.
- 동작 불변 순수 리팩토링. stop_strategy의 remove_category 호출 순서 유지.
- 테스트 ~44곳 receiver를 _event_shadow_manager로 갱신(내부 표면 직접 잠금).
- strategy_scheduler.py 2432→2097줄.

검증: 스케줄러 4개 파일 207 green(전후 동일), unit 6505 + integration 245 green.
